+++ MazeGenerator/MainWindow.xaml.cs
@@ -35,21 +35,12 @@ namespace MazeGenerator
 		{
 			InitializeComponent();
 		}
-		//private void ButtonGenerate_Click(object sender, RoutedEventArgs e)
-		//{
-		//	if (string.IsNullOrEmpty(UpDownHeight.Text) || string.IsNullOrEmpty(UpDownWidth.Text))
-		//		return;
-		//	generator = new EllerAlgorithm((ushort)UpDownHeight.Value, (ushort)UpDownWidth.Value);
-		//	generator.Generate(ref canDoNextStep);
-		//	PrintMaze(sender, e);
-		//	ListBoxPaths.ItemsSource = null;
-		//}
 		ManualResetEvent signal;
 		private void ButtonGenerate_Click(object sender, RoutedEventArgs e)
 		{
 			if (string.IsNullOrEmpty(UpDownHeight.Text) || string.IsNullOrEmpty(UpDownWidth.Text))
 				return;
-			generator = new EllerAlgorithm((ushort)UpDownHeight.Value, (ushort)UpDownWidth.Value);
+			generator = new EllerAlgorithm((int)UpDownHeight.Value, (int)UpDownWidth.Value);
 			CancellationTokenSource cancelSource = new CancellationTokenSource();
 			Progress<string> progress = new Progress<string>(s => OnNextStep(s));
 			ListBoxPaths.ItemsSource = null;
@@ -63,31 +54,6 @@ namespace MazeGenerator
 			if ((bool)CheckBoxSteps.IsChecked || msg == "Search has ended" || msg == "Generated")
 				ImageMaze.Source = generator.ToBitmap(wallPx, cellPx);
 		}
-		//private void ButtonSearch_Click(object sender, RoutedEventArgs e)
-		//{
-		//	if (string.IsNullOrEmpty(UpDownHeight.Text) || string.IsNullOrEmpty(UpDownWidth.Text))
-		//		return;
-		//	if (generator == null)
-		//		ButtonGenerate_Click(sender, e);
-		//	if (!(generator is Searcher))
-		//		generator = new ModifiedDFS(generator);
-		//	Searcher searcher = generator as Searcher;
-		//	searcher.Search(ref canDoNextStep);
-		//	PrintMaze(sender, e);
-		//	ListBoxItem[] items = new ListBoxItem[searcher.paths.Count];
-		//	for (int i = 0; i < items.Length; ++i)
-		//	{
-		//		ListBoxItem tmp = new ListBoxItem();
-		//		CheckBox check = new CheckBox();
-		//		tmp.Margin = new Thickness(2, 2, 2, 0);
-		//		check.Content = $"Path {i + 1}";
-		//		check.IsChecked = true;
-		//		check.Click += ListBoxPathsChange;
-		//		tmp.Content = check;
-		//		items[i] = tmp;
-		//	}
-		//	ListBoxPaths.ItemsSource = items;
-		//}	
 		private void ButtonSearch_Click(object sender, RoutedEventArgs e)
 		{
 			if (!(generator is Searcher))
